+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* ACCERT plugin via the `PluginACCERT` class
+* ACCERT plugin via the `PluginACCERT` class ([#83](https://github.com/watts-dev/watts/pull/83))
 
 ### Changes