+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 * The `Plugin.__call__` method now accepts an `output_dir` argument that
-  specifies the directory created in the database.
+  specifies the directory created in the database
+  ([#107](https://github.com/watts-dev/watts/pull/107))
 
 ## [0.5.1]